Rename common module to employees
